Add for loop building and array indexing
Arrays are now structs behind the scenes to accommodate utilizing their length in various places such as when iterating.

They can be seen as `{ ptr, len }` where `ptr` is a ptr directly to their value. We may add to this behaviour as we determine the best course for dynamic and heterogenous array types in Jink.